Towers that detect enemies
Made towers that spawn at random locations (the spawning mechanism uses random number generators for the x and y components). When an enemy is within its radius, the enemy will be destroyed and a bullet appears where the tower is. When a tower is clicked on, a circle will appear showing the detection radius. 

On state 3, click on 8 to spawn a tower. There is no limit to the spawn amount. Then click on 7 to spawn a path and click 6 to spawn enemies. If there are any towers with a detection radius that overlaps the path, the enemy shall despawn and a bullet shall appear at the location of the tower.

There are no comments, the code is a bit messy and there is a bit of redundant code due to the lack of time but it should work as intended. 

Future plans are to make the bullets move to the enemies and hit them.