Added hud
Added a hud to the game. This includes a menu that allows the player to change scene, a menu to make or remove towers and speed setting buttons. There are also two types of towers, enemies, bullets and paths. The parameters for the different objects like enemies are now stored in a resources scene class which is accessed during play.