Altered HUD and enemies can spawn in waves
The HUD has been changed to show some basic stats like damage and speed of the tower that was clicked on. Hovering the mouse over a tower that could be bought also shows the stats of that specific tower. The game can change speed so the tower fire rate, bullet speed and enemy speed can change on the fly. The game now pauses when the settings window is up. The infrastructure for enemies to spawn in waves has been set up two buttons to make waves with. The spawn rate does not dynamically change with the game speed changes yet. All the image assets have their loading moved to the resources scene so all image references implicitly point to that.