Game Map is Tiles
The background map is now made of tiles instead of images. This means that making maps is made easier. The functional parts of the game like enemies, towers and the HUD are still objects. Instead of following a path made by the developer, the enemies now follow a path made using an external library called 'easystarjs' so this needs to be installed via Node. The package.json has been updated to reflect this change.