Level select
Able to select levels from the level select state (2). For now the levels look similar but have different starting lives and money. All the level, tower, bullet and enemy data are now in JSON files and loaded into the resources state upon loading instead of being in the code for the resources state.