AOE Towers
The ability to make AOE towers has been added to the game. Also, all the importing of files has been moved to the preload of the resources object. This means that there are no more 'import x from y' code at the top. Instead, the names and paths of external files have been fed to a dictionary which is then processed later in the preload.