HP bars and split JSON files
The enemies now have HP bars which hover above the individual enemies and are the same size as the enemy in question. The JSON files are also now split into small files so each one only deals with a single tower, bullet, enemy or level this makes editing different parts of the game easier if need be.